AI-Coding-Assistant/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai_coding_assistant"
version = "0.1.0"
description = "A production-grade AI coding assistant"
readme = "README.md"
authors = [
    {name = "M Sai Sri Datta", email = "msaisridattaDev@gmail.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.8.0",
    "openai>=1.0.0",
    "langchain>=0.1.0",
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "websockets>=11.0.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "faiss-cpu>=1.7.4",
    "chromadb>=0.4.0",
    "gitpython>=3.1.30",
    "tree-sitter>=0.20.0",
    "pygments>=2.15.0",
    "requests>=2.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
    "pre-commit>=3.0.0",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
